+++ ayeyarwady/types.py
@@ -4,4 +4,4 @@ INT8 = ir.IntType(8)
 INT16 = ir.IntType(16)
 INT32 = ir.IntType(32)
 
-FLOAT = ir.FloatType()
+DOUBLE = ir.DoubleType()
